% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Quantiles.R
\name{get.quantiles.recon}
\alias{get.quantiles.recon}
\title{Get prior and posterior quantiles for results of Bayesian population reconstruction.}
\usage{
get.quantiles.recon(param = c("srb", "fert.rate", "mort.rate",
  "mig.count", "baseline.count", "tfr", "e0", "IMR"), results.recon,
  results.post.process.recon, results.prior = NULL,
  results.post.process.prior = NULL)
}
\arguments{
\item{param}{Character vector of parameters for which quantiles are wanted.}

\item{results.recon}{Result of running \code{\link{pop.est.sampler}}.}

\item{results.post.process.recon}{Result of running \code{\link{post.process.recon}} on \code{results.recon}.}

\item{results.prior}{Result of running \code{\link{sample.from.prior}}.}

\item{results.post.process.prior}{Result of running \code{\link{post.process.recon}} on \code{results.prior}.}
}
\value{
Dataframe with prior and posterior marginal quantiles for requested parameters.
}
\description{
Convert outputs of \code{\link{pop.est.sampler}} and
\code{\link{sample.from.prior}} and produce marginal quantiles fo
various parameters, returned in a conventient form for plotting.
}
\author{
Mark C. Wheldon
}
