% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseHyperparameters.R
\name{make.hyper.params}
\alias{make.hyper.params}
\title{Convert elicited relative errors into hyperparameters}
\usage{
make.hyper.params(absDev = list(fert = 0.1, surv = 0.1, mig = 0.2, pop =
  0.1), prob = list(fert = 0.9, surv = 0.9, mig = 0.9, pop = 0.9),
  alpha = list(fert = 0.5, surv = 0.5, mig = 0.5, pop = 0.5), s.star)
}
\arguments{
\item{absDev}{List of the 'p's in the elicitation statement.}

\item{prob}{List giveing probabilities with which the elicitation
statement holds.}

\item{alpha}{List of inverse gamma alpha parameters.}

\item{s.star}{Initial estimates of survival proportions.}
}
\value{
List with elements \code{al.f}, \code{al.be}, etc., the
alpha and beta parameters for each of the components
\dQuote{fert}, \dQuote{surv}, \dQuote{mig}, \dQuote{pop}.
}
\description{
Given elicited relative errors from the statement:
"With probability 'prob', the true values are within 'p' percent of the initial estimates",
generate hyperparameters for Bayesian reconstruction.
}
\author{
Mark C. Wheldon
}
