% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_variances.R
\name{plot.acceptance.props}
\alias{plot.acceptance.props}
\title{Plot acceptance proportions for a Bayesian popualtion reconstruction}
\usage{
\method{plot}{acceptance.props}(results.recon, ylim = c(0, 1))
}
\arguments{
\item{results.recon}{Output from \code{\link{pop.est.sampler}}.}

\item{ylim}{Limits for y axes for plots.}
}
\value{
Nothing. Creates a plot of acceptance proportions.
}
\description{
Metropolis acceptance proportions should be between about 0.2 and
0.5 for 'good' mixing. This function plots them for a completed
run of \code{\link{pop.est.sampler}}. They can be accessed
directly via
\code{\var{results.recon}$alg.stats$acceptance.proportions}.
}
\author{
Mark C. Wheldon
}
