% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstModel_TwoSex.R
\name{pop.est.sampler}
\alias{pop.est.sampler}
\title{Generate MCMC samples for two-sex Bayesian reconstruction}
\usage{
pop.est.sampler(n.iter, burn.in = 0, al.f, be.f, al.s, be.s, al.g, be.g,
  al.n, be.n, al.srb, be.srb, mean.f, mean.s, mean.g, mean.b, mean.srb,
  start.f, start.s, start.g, start.b, start.srb, start.sigmasq.f,
  start.sigmasq.s, start.sigmasq.g, start.sigmasq.n, start.sigmasq.srb,
  pop.data, fert.rows = as.logical(apply(mean.f == 0L, 1, function(z)
  !all(z))), proj.periods = ncol(mean.f), age.size = 5, verb = FALSE,
  progress.step = 1000, prop.varcovar,
  save.mid.run.every = progress.step, save.mid.run.name = NULL)
}
\arguments{
\item{n.iter, burn.in}{Number of saved iterations and burn-in iterations.}

\item{al.f, be.f, al.s, be.s, al.g, be.g, al.n, be.n, al.srb, be.srb}{Fixed hyperparameters (same for male and female). See \code{\link{make.hyper.params}}.}

\item{mean.f, mean.s, mean.g, mean.b, mean.srb}{Fixed prior means. Lists with \code{female} and \code{male} components.}

\item{start.f, start.s, }{start.g, start.b, start.srb Inital values for vitals. Not transformed coming in. Lists with \code{female} and \code{male} components.}

\item{start.sigmasq.f, start.sigmasq.s, start.sigmasq.g, start.sigmasq.n, start.sigmasq.srb}{Inital values for variances (same for male and female).}

\item{pop.data}{Census data. Not transformed coming in. List with \code{female} and \code{male} components.}

\item{fert.rows}{Rows of \code{mean.f} and \code{start.f} that are non-zero.}

\item{proj.periods}{Number of periods to project forward over (e.g., number of five-year steps).}

\item{age.size}{Age group width.}

\item{verb, progress.step}{print algorithm progress every 'progress.step' iterations}

\item{prop.varcovar}{Variances for proposal distributions used in M-H steps which update vital rates. Lists with \dQuote{female} and \dQuote{male} components.}

\item{save.mid.run.every, save.mid.run.name}{How often should the chain be saved, and with what name?}
}
\value{
List with components for CCMPP and other algorithm parameters.
}
\description{
The main function that does the estimation of parameters for the reconstruction.
}
\author{
Mark C. Wheldon
}
