% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Counts.R
\name{post.process.recon}
\alias{post.process.recon}
\title{Convert CCMPP parameters (rates, proportions, etc.) into counts}
\usage{
post.process.recon(results.recon, sep.factors = list(female = 0.5, male =
  0.5), outputs = c("birth.count", "total.birth.count",
  "total.birth.death.count", "death.count", "mig.count", "total.mig.count",
  "mort.rate", "mig.rate", "person.years", "cohort.nq0", "period.nq0",
  "IMR"), name.pref = character(0), name.suf = character(0))
}
\arguments{
\item{results.recon}{List with components \code{fert.rate.mcmc},
\code{surv.prop.mcmc}, \code{mig.prop.mcmc},
\code{baseline.count.mcmc}, \code{pop.count.mcmc},
\code{srb.mcmc}. All but the last are lists of mcmc objects with two
components named 'female' and 'male'. Column names are
yyyy.aa.aa (year.age-start.age-end). \code{mig.prop.mcmc} is
assumed to be the average \emph{annual} proportion. It is
multiplied by age.int to get migration for the entire
interval. \code{pop.count.mcmc} does not contain
baseline.count.mcmc. These should all have the same number of
rows. The output will have this many rows too, otherwise the
same number of rows as the smallest of these inputs.

\code{srb.mcmc} is a numeric vector, length 1, or an mcmc/matrix
with one column per period (usually 5-year period).}

\item{sep.factors}{List of separation factors used to separate
cohort deaths into period deaths with components names
'female' and 'male. These can be a scalar, in which case the
same factor is used for all ages, year, or a \emph{vector} with
names yyyy.aa indicating the 'y'ear and 'a'ge. Age goes from 0
to the open ended interval in the population counts (not the
survival proportions).

For sep factor k_a, {delta}_D_a^c = k_a * D_a^c, i.e., k_a is the
fraction of cohort deaths for cohort aged 'a' that occurs in the
first triangle of the cohort parallelogram. sep factor for oldest
age group is currently re-set to 1.}

\item{outputs}{Character vector listing the counts required as
output. Counts are calculated regardless but are not output
unless requested.}

\item{name.pref}{String to prepend to output objects.}

\item{name.suf}{String to append to output objects.}
}
\value{
The requested outputs are matrices and are placed in the
Global environment.
}
\description{
Converts the quantities used in population projection (fert rates
suvival proportions, migration proportions, counts) into counts of
births, deaths and net number of migrants.
}
\author{
Mark C. Wheldon
}
