% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFromPrior.R
\name{sample.from.prior}
\alias{sample.from.prior}
\title{Draw a random sample from the joint prior distribution}
\usage{
sample.from.prior(n.iter = 1000, batch.size = ceiling(n.iter/4), al.f,
  be.f, al.s, be.s, al.g, be.g, al.n, be.n, al.srb, be.srb, mean.f, mean.s,
  mean.g, mean.b, mean.srb, fert.rows = as.logical(apply(mean.f == 0L, 1,
  function(z) !all(z))), proj.periods = ncol(mean.f), name.pref = "",
  name.suff = "", age.size = 5, message.every = 1000,
  neg.pop.tol = 0, parallelize = TRUE, cores = NULL,
  pvm.hostfile = "~/.pvm_hosts")
}
\arguments{
\item{n.iter}{Size of sample to draw.}

\item{batch.size}{Size of batches in which to make draws.}

\item{al.f, be.f, al.s, be.s, al.g, be.g, al.n, be.n, al.srb, be.srb, mean.f, mean.s, mean.g, mean.b, mean.srb, start.f, start.s, start.g, start.b, start.srb, start.sigmasq.f, start.sigmasq.s, start.sigmasq.g, start.sigmasq.n, start.sigmasq.srb, fert.rows, proj.periods, age.size}{See \code{\link{pop.est.sampler}}.}

\item{message.every}{Control progress reporting; print a message every \code{message.every} iterations.}

\item{parallelize}{Run in parallel?}

\item{cores}{Number of parallel processes (cores) to use if \code{parallelize} is true.}

\item{pvm.hostfile}{pvm hosts for parallelization via SNOW on *NIX.}
}
\description{
Sample from the prior to determine distributions of transformed
parameters such as mortality rate and migration counts.
}
\author{
Mark C. Wheldon
}
