\docType{data}
\name{Thailand_Example}
\alias{Thailand_Example}
\title{Thailand Example.}
\format{RData archive with the following objects:
   \describe{
   \item{asFertTHAI.mat}{Matrix of initial estimates for age-specific (female) fertility rates.}
   \item{asSurvTHAI.mat}{List of matrices of initial estimates for age-specific survival proportions.}
   \item{asMigTHAI.mat}{List of matrices of initial estimates for age-specific migration proportions.}
   \item{asbaselineTHAI.mat}{List of matrices of initial estimates for population counts in the baseline year.}
   \item{censusTHAI.mat}{List of matrices of bias-adjusted population counts (based on censuses).}
   \item{srbTHAI.mat}{Matrix of initial estimates of sex ratios at birth.}
   \item{thai.propvar}{Proposal variances for Metropolis-Hastings proposal distributions.}
   \item{thaiFemale.sf}{Life table separation factors for females.}
   \item{thaiMale.sf}{Life table separation factors for males.}
 }}
\usage{
data("Thailand_Example")
}
\keyword{datasets}




